refactor: new comment modal
